LOOK MAN IT AINT DONE BUT I SURE AS HECK DONT WANNA BE BORED

+++ src/cap_math.c
@@ -2,15 +2,15 @@
 
 float lerp(float a, float b, float f)
 {
-    return a * (1.0 - f) + (b * f);
+    return a * (1.0f - f) + (b * f);
 }
 
 float degToRad(float ipt)
 {
-    return ipt * M_PI / 180.f;
+    return ipt * (float)M_PI / 180.f;
 }
 
 float radToDeg(float ipt)
 {
-    return ipt * 180.f / M_PI;
+    return ipt * 180.f / (float)M_PI;
 }
